Add support for MSIAvailable on older versions of Go
http.NewRequestWithContext() was added in Go 1.13.  This change uses
build constraints to support older versions.
Note that Go 1.7 is the minimum version due to reliance on the context
package.
